Change the internal storage to Vec
Prepares reallocation and resizing. It is also unecessary to
convert to a boxed slice and this potentially reallocates to remove
overallocated capacity (not currently due to the macro used for
construction but it would be easy to do so accidentally).

Additional remark: The test runs fine with current nightly and `miri`,
this should stay that way when the test suite is expanded.