Provide a dynamically sized aligned slice
Changes the Buf to a Buffer that works similar to a Vec, and provides a
DST buf that provides what slices are to Vec. Previous instance methods
on the Buffer are now provided on the slice and a Deref(Mut) impl
ensures that they are still provided by the Buffer as well.