Keep track of byte length
Similar to `Vec` we keep track of the logical length of initialized
memory with a separate member in Rec. This makes the apparent effects of
many operations more reasonable, as input parameter match the resulting
properties. Previously, a `Vec` initialized with e.g. 7 bytes would
result in a byte_len of 16, unecessarily exposing the properties of the
underlying chunk size.