Redefine internal Buffer reallocation usage
Clarifies that its use is not to provide an exact allocation length but
just an aligned byte buffer with some level of control over the number
of elements. It is to `Rec` similar as `RawVec` is to `Vec` in the
standard library. Keeping track of logical length is now (correctly) in
complete control of the higher level `Rec`.

The effect is that a sane definition of `reuse` for resizing without
reallocation is possible.