Reallocation for canvas and improved errors
Buffer reuse for canvas provides two interfaces: One for reusing the
inner buffer during a layout change and one for reusing an existing
buffer while construction a canvas. The latter ensures that the buffer
itself remains in the control of the programmer by returning it as part
of the error. This effecitvely also enables recasting the image for a
different pixel type with different size by converting to a raw Rec
buffer first, reinterpreting that, and then creating another Canvas.
All of these operations will preserve any existing image byte data.