Add more standard trait derives
Implement comparisons, Debug, Clone, and Default for many surface types.
The annoying part is that they can not be derived by default for types
that merely use `Pixel<P>` but do not actually contain an instance of
`P` itself. The auto-derive has an unecessary bound on the inner type
which is not required similar to PhantomData.