Switch to bytemuck
Pros:
The crate is more minimal
The source code comes with a public bug tracker
The crate has a useful stable version
The minimum Rust version is 1.34

Cons:
We lose the distinction of AsBytes and FromBytes. But firstly, this
might be temporary. Depending on how bytemuck will procede there might
be a chance to get this back. Secondly, this has not really proven
useful. The only real chance to use it would have been to somehow
allocate the data yourself which had not been supported.

All pixel types are required to be Copy. That's probably good as no
longer is there any unexpected not-occurring Drop. The bytemuck crate
seems more reasonable here.