Hide implementation detail of using buf better
Interfaces that are copy-on-write (or copy-on-grow here) are not very
typical for the Rust ecosystem. Rather, add the buffer type as an
additional type parameter. This would however leak the implementation
detail of `Cog` and `buf`. To avoid this we introduce one additional
crate-private inner type to hold all implementations but which has such
a type parameter. Then all implementations can make direct use of the
type—such as Matrix that can instantiate it with an owned Buffer now.
This will then only require 3-4 wrappers to expose the different
borrowing modes to the world. Maybe we'll even drop `Cog` in time.