A very basic concept of raster layouts
The idea here is to separate it from the actual buffer and apply one for
layouts. We avoid references here and only deal with value types to
enable planar raster images. These can insert and retrieve full pixels
by treating each channel on its own but wouldn't be able to construct a
single reference overing the pixel itself, as such a thing does not
exist.

For performance we might later extend this with default implemented
methods that are allowed to edit pixels in-place in layouts that support
such edits.