Add a byte matrix layout
Used for copying raw byte data, as efficiently as possible in the
future. The main trick would be that most other matrix layouts are
convertible to it and hence do not need to implement their own copying.

Additionally, inlining or code flow would choose efficient memcpy based
operations on whole contiguous rows and columns [WIP].