Define color conversion representation
We use a macro to avoid a generic type in any of the interface. The
implementation, however, will not use macros, at least in the current
plan. Instead, matching and match arms would ensure that no case is
missed and allows special casing of individual types where efficiency
can be improved.