Fixup and speedup rgba conversion
The mask applied to each pixel in the shuffle part was incorrect. Oops.
This allows us do simple reshuffles directly on each texel without
expanding them into common color / float and having to recast down
again.

We could save the copy to the texel buffer in the first place as well.