Speedup coordinate generation
This is still wasteful, and takes 26% of time or so.. We should just
specify *what* we want to do (i.e. (x, y) and count) as a special case.
In particular, layout's method can totally special case on this to
generate everything with SIMD. Or even not generate anything because
we'll use the underlying chunk if possible.