Refactor image into raw inner and public interface
Similar to Box and its inner unique in the standard library, this should
allow better sharing of implementations between buffer types where
otherwise we would leak the rather private traits that contain the
shared behaviour. And we only want that trait to make sense for our own
selection of buffers since the exposed image containers are all concrete
buffers.