Relax bounds on raw buffer, avoid Deref to buf
The largest use of that deref is the methods available on slices and on
`&buf` such as aligned operations and lengths. These can be provided in
a much more generic way even for underlying representations that are not
that specific (`Freeze`) buffer type but also for !Sync byte buffers and
atomic buffers. We just need to contract the composed operations for
modifying such buffers so that the signatures only mention the Self type
of them respectively. These additional methods will be added to
`BufferLike` in time.