Improve stores into atomic slice
Firstly, reduces the amount of soundness documentation required by
making different data sources distinct. (The loads are funnily asymetric
here since the output mutable reference is very strong and can be easily
converted to cells, but there is not opposite conversion to a shared
reference).

Secondly, optimizes the 'bulk' of a long atomic store operation by doing
the interior in a single store without any of the xor dance. Here we
modify all bytes so none of the previous value needs to be preserved.
Only when we straddle the edge to some neighbor values that should not
be affected by our modifications do we need to do the modifications by
xor to mask to the bits we want to affect.