Polish decay based interfaces
In particular we expect these to be infallible by default, the check on
length is redundant if implementations perform according to the trait
requirements. Thus, we design the APIs around returning a definitive
result.

This allow using decay as an unobstructive mechanism for 'do not care
about layout' through the fact that everything decays to `Bytes`.