Move Yuv420 layout into canvas, drop DynLayout
These is really quite a bit more specialized. The Yuv420 plane encoding
in particular is a modification of a standard Yuv which is a part of the
exhaustive (not open, but growing) list of ITU recommendations. It fits
better into the opinionated crate of `canvas` than the generalist
`texel` due to this reference.

DynLayout was ill-specified and therefore did not get any attention for
a long time. It is simply dropped. In `canvas` we will define a much
more well-defined variant of it by extending its nominal `Layout` type
with such capabilities based on the color and bit format.

This completes the work of removing superfluous stuff.