Inherent non-bytemuck covered Texels now unsafe
This changes the internal macro defining the inherently available
`Texel` implementations. The macro calls now take an additional unsafe
keyword and comment to justify that usage. When provided the call allows
arbitrary types but otherwise requires the type to be `bytemuck::Pod` at
const time. This works around having to surface all kinds of different
features covering parts of bytemuck's SIMD implementations.

I expect that some of this will become less intricate if
`#[cfg(rust-version())]` (or w/e it is spelled) will be supported since
this would get rid of the need for such feature flags in bytemuck. Iff
they bump to `2` as to only support compiler versions that support such
a configuration macro. Which.. yeah. Anyways we can at least add
implementations for new SIMD types as they are stabilized.