Test but disable LumaDigital color
We have inadequate test coverage for these. There is a standard
somewhere giving us a description and maybe even test vectors but this
also implies the definition may be bogus. For instance, I'm not certain
if the Whitepoint/Rgb gives the weighting coefficients directly or if
diverging values are chosen to compensate and approximate for the messed
up weighting of dark channels at high ''gamma'' transfers.