Qualify RGB conversions with colour-science
Previous attempts at reproducing the numbers failed to use the argument
for marking inputs such that the transfer function must be applied,
which is *not* the default despite what you might expect when hearing
sRGB and similar. Thus the conversion matrices obviously differed.

This reproduces floats up to 3 digits which is .. not quite as good as I
had expected but I think the Python library is using f64 for everything
where our matrices are still f32 calculations and such intermediate
results. A look at using more precise and faster matrix multiplication
with explicit instructions should be done at a later point.