Add unaligned byte planes to image-canvas
These interact with the shader / conversion machinery. We can load any
texel even from an unaligned location, not optimally but of course it
does work for all Pod types. Notably these are not optimized for simple
shuffles, yet, despite intrinsics spending most of their work in the
shuffle instructions and only little bit in what would be loadu and
storeu.